changed return type

+++ src/utils/time_utils.py
@@ -77,7 +77,7 @@ class TimeUtils:
         return None
 
     @staticmethod
-    def get_gap_duration(segment_1: Segment, segment_2: Segment) -> int:
+    def get_gap_duration(segment_1: Segment, segment_2: Segment) -> float:
         """
         Calculates gap between two segments (the gap between start of the second segment and end of the first segment.)
         Args:
@@ -88,4 +88,4 @@ class TimeUtils:
             int: Rounded down number of seconds.
         """
         difference_in_ms = segment_2.total_ms_start - segment_1.total_ms_end
-        return int(difference_in_ms // 1000)
+        return difference_in_ms // 1000
